Make the DOMException constructor accept a name
It will automatically set the code from the name (if it is one of the legacy names that have codes).

It now just calls `super(message)` to initialize the message and all of the internal error-related state.

The "creating exceptions" section was refactored to work in terms of the new, fully-functional constructor.

Closes https://www.w3.org/Bugs/Public/show_bug.cgi?id=27062.