Remove mention of OMG IDL
We haven't used OMG IDL on the web in over a decade (and rapidly approaching two decades). Might be time to erase it from our collective memory. One should not be required to know anything about OMG IDL read this spec, and current spec mentioning it seems to imply otherwise.