Reword the distinguishability algorithm.
This slightly clarifies the wording saying that nullable and dictionary
types aren't distinguishable, and adds an example to make it painfully
clear.

It also handles nullable unions before the recursive cases strip
nullability so we handle the case of "nullable union and union containing a
nullable type" correctly.

It also adds some examples for applying the distinguishability table.

Fixes #193