Re-align DOMException objects with what is implemented
This fixes #55. As discussed there, the definition of DOMException does not match any implementations. https://github.com/heycam/webidl/commit/137856d678f67b6f79f0c3ef4bd7d233629a685f exacerbated this, but it was already incorrect (e.g. defining own properties). Instead, implementations seem to implement DOMException as an interface with a small amount of custom bindings. This updates the specification to reflect that.

As part of this, it is no longer a distinct type, but instead an interface type.