Restrict the extended attributes for partial interfaces.
It seems like a bad idea to allow [Global] on a partial interface.

The behavior of [Exposed] and [SecureContext] on partial interfaces is already
defined in the relevant algorithms.

The behavior of [OverrideBuiltins] is defined here with the same vague wording
used before. It might be better to disentangle the concept of an interface and
its syntactical structure.

Fixes #154.