Correct requirement to define Global properties on the object itself
The specification required defining the operations and attributes for the Global
interface itself as well as its inherited interfaces on the object itself.

This is not implemented in Gecko, Chrome or WebKit: none of these engines
defines, for example, the `addEventListener` method on the window object.

I suspect this text was meant to refer to mixin interfaces rather than
inherited interfaces, but these are already covered by a catch-all statement.