Remove support for overloading callbacks (#690)
This was only supported in the definition of the effective overload set, but
this wasn't actually used anywhere, and there is no syntax to overload a
callback anyway.

The six types of IDL constructs are the five that were in the specification
before this change, and legacycallers. The count was not updated when these
were removed.