Add [AllowShared] to the list of applicable-to-types
Annotated types lists 3 extended attributes as applicable-to-type attributes, and [AllowShared] is not in the list. On the other hand, [AllowShared] is explained to appear on one of the buffer source type, and it means [AllowShared] is applicable to types.

This adds [AllowShared] in the list of applicable to types extended attributes, so that it fixes the ambiguity.

Fixes #687.