Remove a check for array index in [[Set]]
Array index properties are not special on objects with named setters. Note
that the corresponding step in [[DefineOwnProperty]] does not have this
condition.

This change does not affect normal usage; it is only detectable by messing with
the prototype of the object.

Gecko and WebKit already handle array indexes the same way as other properties;
Chrome does for DOMStringMap but does not for Storage. (In the Storage case, it
does the exact opposite of the current specification.)