Editorial: make more iterator-related stuff imperative
As part of #467, this moves the installation of @@iterator/@@asyncIterator/forEach/entries/keys/values to an imperative style, for the cases of pair iterators, value iterators, async iterators, and indexed property getters.

This does not yet make imperative the same methods for maplikes and setlikes, nor does it make the construction of the iterator-related objects and prototype objects themselves imperative.