Group and alphabetize legacy extended attributes
This groups all the legacy extended attributes under a new section, that follows the general extended attributes section. (It also renames that section from "ECMAScript-specific extended attributes" to "Extended attributes", since we are already in the ECMAScript binding section.)

As part of this, it removes all the individual "do not use this" warnings, in favor of a blanket one at the start of the section. Some cases, namely [LegacyFactoryFunction], [LegacyLenientSetter], and [LegacyNamespace], contained advice or background information in those blocks; those have been preserved.