Give SharedArrayBuffer a dedicated type
Generally IDL objects have been 1:1 with ECMAScript objects, except for
SharedArrayBuffer. That was instead represented as
[AllowShared] ArrayBuffer, although that includes ArrayBuffer as well.
This created a number of challenges, e.g., how to return a
SharedArrayBuffer object.

So we make these changes here that will also require corresponding
downstream fixes:

* SharedArrayBuffer is now its own type.
* [AllowShared] only applies to buffer view types.
* AllowSharedBufferSource takes over from [AllowShared] BufferSource.

Downstream is tracked by #1320.

Fixes #865 and fixes #961.