Introduce a "detachable" algorithm for BufferSource
The new "detachable" algorithm allows an invoker to determine if a BufferSource is detachable
without errors before invoking transfer. This is particularly helpful if several buffers are
processed in a batch, as the validity of the operation can be determined up front, rather than
during processing which could leave the buffers in a mix of states.

See discussion in [1]. A local copy of the algorithm was landed as [2] but this seems like a useful
addition to WebIDL.

1: https://github.com/webmachinelearning/webnn/issues/351
2: https://webmachinelearning.github.io/webnn/#buffersource-detachable