Editorial: Add a Web Platform Design Principles link for `[Exposed=*]`
In w3ctag/design-principles#510, the TAG's Design Principles document
added a section on considerations for exposing an interface everywhere
(`[Exposed=*]`). This patch links to that section in the definition of
`[Exposed=*]`.

Additionally, the paragraph where that link is added used to have
`class="advisement"`, which has no styles or any apparent special
meaning. This patch changes it to `class="note"`, which was probably
the intention.